+++ src/engine/canvas/CanvasLifecycleManager.ts
