+++ src/engine/canvas/__test__/CanvasLifecycleManager.test.ts
